// src/app/components/checkout-page/model/item.model.ts
export interface ItemModel {
  showContactInfo: boolean;
  real_delivery_time: any;
  seller: any;
  result: any;
  data: any;
  _id: string;
  name: string;
  description: string;
  price : any;
  image: string;
  zip: Date,
  country: string,
  city: string;
  street: string;
  number: 2,
  publication_date: Date,
  quantity: string,
  mobile:any;
  codiceFiscale: any;
  buyerEmail: any;
  servicecharge: Number;
  total_price :any;
}